feat(fs): implement tmpfs write with 0/1/2-level block mapping
- Add tmpfs_bmap: resolve/allocate data pages via direct, single- and
  double-indirect blocks using TMPFS_* macros (uint64 slots, 512/page)
- Add tmpfs_vfs_write: page-wise copy under inode lock, track size,
  return bytes written; wire into tmpfs_file_ops
- Add 11 write tests: basic, cross-page, offset, level1/2, invalid args,
  size-max, fill-direct, 0->1 and 1->2 boundaries, fill-indirect
- Fix easyfs off+size overflow cast

+++ kernel/fs/easyfs/fs.c
@@ -91,7 +91,7 @@ int easyfs_write_inode(struct vfs_inode *ip, int user_src, uint64 src,
 
 	if (off + size < off)
 		return -1;
-	if (off + size > MAXFILE * BSIZE)
+	if ((uint64) off + size > MAXFILE * BSIZE)
 		return -1;
 
 	for (tot = 0; tot < size; tot += m, off += m, src += m) {
